PgSTAC: API hydration of search result items (#397)
* Upgrade to pgstac 0.5.1

Initial changes to get most tests passing.

* Add option to hydrate pgstac search results in API

* Support fields extension in nohydrate mode

* Updates to hydrate and filter functionality.

This was done in a pairing session with @mmcfarland

* Fix fields extensions and reduce number of loops

* Tolerate missing required attributes with fields extension

Use of the fields extension can result in the return of invalid stac
items if excludes is used on required attributes. When injecting item
links, don't attempt to build links for which needed attributes aren't
available. When API Hydrate is enabled, the required attributes are
preserved prior to filtering and are used in the link generation.

* Run pgstac tests in db and api hydrate mode

* Merge dicts within lists during hydration

In practice, an asset on a base_item and an item may have mergable
dicts (ie, raster bands).

* Add note on settings in readme

* Pass request to base_item_cache

This will be used by implementors who need app state which is stored on
request.

* Upgrade pypgstac and use included hydrate function

The hydrate function was improved and moved to pypgstac so it could be
used in other projects outside of stac-fastapi. It was developed with a
corresponding dehydrate function to ensure parity between the two.

The version of pypgstac is unpublished and pinned to a draft commit at
the point and will be upgraded subsequently.

* Improve fields extension implementation

Correctly supports deeply nested property keys in both include and
exclude, as well as improves variable naming, comments, and test cases.

* Remove unused error type

* adjust tests for changes in api

* remove print statements

* add bbox back to items in tests

* Upgrade pgstac

* Fix conformance test fixtures

* Fix sqlalchemy test with new status for FK error

* Align fields ext behavior for invalid includes

* Lint

* Changelog

* Remove psycopg install dependency

* Relax dependency version of pgstac to 0.6.* series

* Update dev environment to pgstac 0.6.2

* Changelog fix

Co-authored-by: Rob Emanuele <rdemanuele@gmail.com>
Co-authored-by: Ubuntu <planetarycomputer@pct-bitner-vm.kko0dpzi4g3udak2ovyb5nsdte.ax.internal.cloudapp.net>