Support creating ItemCollections in Transaction Extension (#35)
* Align dev container app path with image

The image copies the root project to /app and matching that through
docker-compose allows host file updates to be copied without rebuilding
the image.

* Implement POST ItemCollection for Transaction Ext

Updates to the transaction spec indicate POST against items should allow
ItemCollection.

* Add consistent validation for item and collection

In the transaction extension, Items and Collection can't have mismatched
ids from the path, but should have the path collection id applied if it
is missing. Ids for both are also not allowed to be a "percent encoded"
value per RFC 3986.

* Add transaction tests

* Changelog

* Align respose type for create Item/ItemCollection

The spec leaves the return type for Item creation open to the
implementation. This change unifies the response of both
Item/ItemCollection POST requests to return an empty response with a
Location header for the newly created single Item, in that case.

* Allow override of valid item/collection ids

Use a setting value instead of a constant so that IDs could be set per
instance.

* Walk back some response unification

After realizing the extent of the breaking change resulting from a
unified response between Item/ItemCollection Tx endpoint, restoring the
original behavior.

* Upgdate dev pgstac version

* isort lint

* deps: soft pin the stac-fastapi versions

---------

Co-authored-by: Pete Gadomski <pete.gadomski@gmail.com>